+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Please refer to [UPGRADING](UPGRADING.md) guide for upgrading to a major version
 
 - Remove deprecation for the raw cURL `CURLOPT_CERTINFO` option
 - Warn when a cURL multi option cannot be applied
+- Clarify that NTLM authentication is deprecated by both Guzzle and curl/libcurl
 
 
 ## 7.13.2 - Upcoming